+++ spec/functional/command/spec_spec.rb
@@ -226,6 +226,14 @@ module Pod
         end
       end
 
+      it 'analyzes the current working directory using Debug configuration' do
+        Dir.chdir(fixture('spec-repos') + 'trunk/Specs/1/3/f/JSONKit/1.4/') do
+          cmd = command('spec', 'lint', '--quick', '--allow-warnings', '--configuration=Debug', '--analyze')
+          cmd.run
+          UI.output.should.include 'passed validation'
+        end
+      end
+
       it 'fails with an informative error when downloading the podspec 404s' do
         WebMock.stub_request(:get, 'https://no.such.domain/404').
           to_return(:status => 404, :body => '', :headers => {})
@@ -328,6 +336,12 @@ module Pod
         UI.output.should.include text.gsub(/\n/, '')
       end
 
+      it 'prints the path of a given podspec respecting the version' do
+        lambda { command('spec', 'which', '--version=3.1.0', 'AFNetworking').run }.should.not.raise
+        text = '3.1.0/AFNetworking.podspec'
+        UI.output.should.include text.gsub(/\n/, '')
+      end
+
       describe_regex_support('which')
     end
 
@@ -348,6 +362,11 @@ module Pod
         output = UI.output.gsub(/[0-9]+: /, '')
         first_spec_path = output.split("\n")[0]
         UI.output.should.include Pathname.new(first_spec_path).read
+      end
+
+      it 'cats the first podspec from all podspecs matching the version' do
+        lambda { command('spec', 'cat', '--version=3.1.0', 'AFNetworking').run }.should.not.raise
+        UI.output.should.include fixture('spec-repos/trunk/Specs/a/7/5/AFNetworking/3.1.0/AFNetworking.podspec.json').read
       end
 
       describe_regex_support('cat')